Support using different columns for filtering
For the InvenTree plugin enable the use of he IPN column for
filtering. Implement caching of IPN to PK. This is somewhat
primitive and assumes that KiCad will always do a select without
filtering before doing a select with filtering.

Update README to use IPN as the key and add notes on not allowing
multiple parts to have the same IPN.